requirements.txt
gradio
pillow
numpy
opencv-python
matplotlib
scikit-learn

